--- benchmark/web/wtf.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head lang="en">
-    <meta charset="UTF-8">
-    <title>WTF Test page</title>
-    <script src="wtf.dart" type="application/dart"></script>
-    <script src="packages/browser/dart.js"></script>
-    <script src="packages/browser/interop.js"></script>
-</head>
-<body>
-</body>
-</html>
