+++ lib/angular.dart
@@ -5,7 +5,6 @@ export 'package:angular/application.dart';
 export 'package:angular/core/module.dart';
 export 'package:angular/directive/module.dart';
 export 'package:angular/core/annotation.dart';
-export 'package:angular/tracing.dart';
 export 'package:angular/introspection.dart' hide
     elementExpando, publishToJavaScript;
 export 'package:angular/formatter/module.dart';
