+++ pubspec.yaml
@@ -24,6 +24,7 @@ dependencies:
   intl: '>=0.8.7 <0.12.0'
   perf_api: '>=0.0.8 <0.1.0'
   route_hierarchical: '>=0.4.22 <0.5.0'
+  tracing: '>=0.1.0 <0.2.0'
 dev_dependencies:
   benchmark_harness: '>=1.0.0 <2.0.0'
   guinness: '>=0.1.9 <0.2.0'
