+++ test/angular_spec.dart
@@ -247,15 +247,6 @@ main() {
         "angular.routing.RouteViewFactory",
         "angular.routing.RoutingModule",
         "angular.routing.ngRoute",
-        "angular.tracing.createScope",
-        "angular.tracing.detectWTF",
-        "angular.tracing.endAsync",
-        "angular.tracing.enter",
-        "angular.tracing.enter1",
-        "angular.tracing.leave",
-        "angular.tracing.leaveVal",
-        "angular.tracing.startAsync",
-        "angular.tracing.wtfEnabled",
         "angular.watch_group.PrototypeMap",
         "angular.watch_group.ReactionFn",
         "angular.watch_group.Watch",
